backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
torch==2.3.0
torchvision==0.18.0
torchaudio==2.3.0
timm==1.0.3
numpy==1.26.4
Pillow==10.3.0
opencv-python-headless==4.9.0.80
SQLAlchemy==2.0.30
alembic==1.13.1
aiosqlite==0.20.0
pydantic==2.7.1
pydantic-settings==2.3.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.27.0
celery==5.4.0
redis==5.0.6
boto3==1.34.120
python-dotenv==1.0.1
loguru==0.7.2
av==12.0.0
scikit-learn==1.5.0
